backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pypdfium2
pybind11==2.12.0
zxing-cpp==2.3.0
pydantic==2.6.1
prometheus-client==0.19.0
python-multipart==0.0.9
PyYAML==6.0.1
pydantic-settings
Pillow 
